refactor(plugin-sdk): share boolean action param parsing
(cherry picked from commit 741e74972b820cf1c42b241450d290c00c6e9e73)
