+++ src/infra/outbound/message-action-params.ts
@@ -9,29 +9,12 @@ import type {
 } from "../../channels/plugins/types.js";
 import type { RemoteClawConfig } from "../../config/config.js";
 import { extensionForMime } from "../../media/mime.js";
+import { readBooleanParam as readBooleanParamShared } from "../../plugin-sdk/boolean-param.js";
 import { parseSlackTarget } from "../../slack/targets.js";
 import { parseTelegramTarget } from "../../telegram/targets.js";
 import { loadWebMedia } from "../../web/media.js";
 
-export function readBooleanParam(
-  params: Record<string, unknown>,
-  key: string,
-): boolean | undefined {
-  const raw = params[key];
-  if (typeof raw === "boolean") {
-    return raw;
-  }
-  if (typeof raw === "string") {
-    const trimmed = raw.trim().toLowerCase();
-    if (trimmed === "true") {
-      return true;
-    }
-    if (trimmed === "false") {
-      return false;
-    }
-  }
-  return undefined;
-}
+export const readBooleanParam = readBooleanParamShared;
 
 export function resolveSlackAutoThreadId(params: {
   to: string;
